refactor: extract search shared types
