+++ Cargo.toml
@@ -239,6 +239,7 @@ members = [
     "src/api_query",
     "src/api_common",
     "src/core",
+    "src/search",
     "src/jobs",
     "src/config",
     "src/infra",
@@ -268,6 +269,7 @@ flight = { path = "src/flight" }
 tantivy_utils = { path = "src/tantivy_utils" }
 web = { path = "src/web" }
 openobserve-core = { path = "src/core" }
+openobserve-search = { path = "src/search" }
 openobserve-api = { path = "src/api" }
 openobserve-api-management = { path = "src/api_management" }
 openobserve-api-query = { path = "src/api_query" }
