+++ src/search/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "openobserve-search"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+publish = false
+
+[lib]
+path = "src/lib.rs"
+
+[features]
+default = []
+enterprise = []
+
+[dependencies]
+config.workspace = true
+serde.workspace = true
+utoipa.workspace = true
+
+[dev-dependencies]
+arrow.workspace = true
+arrow-schema.workspace = true
+o2_enterprise.workspace = true
+serde_json.workspace = true
